revert: remove  emoji support in lexer and update NoHashEmojiRule (#68)
Previously, we extended the lexer to parse the  emoji for attribute declarations.
However, PHP does not support it for attributes, so we are reverting that lexer change.
We also updated `NoHashEmojiRule` to clearly warn that using  in attributes
doesn’t parse at all, while still discouraging its use in comments for clarity and tooling compatibility.

Signed-off-by: azjezz <azjezz@protonmail.com>