feat(cli): add `—workspace` option and update configuration
This commit adds the `—workspace` option to the Mago CLI, allowing users to specify the workspace directory for their project. This option also controls where the configuration file is loaded from.

The `source.root` configuration option has been renamed to `source.workspace` to better reflect its purpose.

Additionally, the documentation has been updated to reflect these changes and to clarify the behavior of the `—workspace` option and the `source.workspace` setting.

Signed-off-by: azjezz <azjezz@protonmail.com>