feat(lint): add --format/--fmt option to 'mago lint --fix'
This commit introduces the `--format` (or `--fmt`) option to the `mago lint --fix` command. When used, it automatically formats the files after linting fixes are applied, eliminating the need for a separate  command.

This enhancement streamlines the linting and formatting workflow, providing a more convenient and efficient experience for users.

Signed-off-by: azjezz <azjezz@protonmail.com>