feat(analyzer): support instanceof with "complex" expressions
Previously, type assertions for `if ($x instanceof $y)` were only generated when `$y` was a simple identifier (`self`, `parent`, `static`, or a class name).

This change extends support to cases where `$y` is a complex expression. The analyzer now evaluates the RHS expression to determine its possible class name types.
This enables powerful type inference for dynamic `instanceof` checks, particularly with templates.

For example, given a variable `$class_string` with the type `class-string<T>`, the following code is now correctly analyzed:

```php
if ($value instanceof $class_string) {
    // $value is now correctly inferred as an instance of `T`,
    // whereas it was previously `mixed`.
}
```

Additionally, a new safeguard has been added. If the type of the RHS expression cannot be determined, we now conservatively assert that the LHS variable is at least an object,
which is the widest possible type it can be within the if block.

Signed-off-by: azjezz <azjezz@protonmail.com>