feat(parser, cli, stubs): add PHP `8.5` support and adopt rolling version policy
This commit introduces support for upcoming PHP features and establishes a new versioning strategy for Mago to always
stay on the cutting edge.

The parser has been updated to support clone as a first-class function (`clone()`), including the new `$withProperties` argument for the "clone-with" feature implemented in PHP 8.5.
This required treating clone as a soft reserved keyword to allow it to be parsed as both an expression and a function call.

Mago will now default to the latest stable PHP version (`LATEST`) and support analysis up to the next minor version (`NEXT`).
This ensures that users can always analyze code using features from the next PHP release ahead of time. With this change, Mago now fully supports PHP `8.5`.

Stubs have been added for a wide range of new functions, classes, constants, and methods from PHP 8.5, including:

- New functions like `get_error_handler()`, `array_first()`, and `curl_multi_get_handles()`
- The new `IntlListFormatter` class and additional Locale methods
- New constant `PHP_BUILD_DATE`
- Improved conditional return types for `array_key_first()` and `array_key_last()`

Signed-off-by: azjezz <azjezz@protonmail.com>