fix(analyzer): resolve generic parameters in `@return static`
Fixes an issue where the analyzer would ignore generic type parameters provided in `@return static<...>` docblocks.
This led to incorrect type inference for methods that transform a collection's generic types, such as map.

This was caused by two problems: the type expander was incorrectly overwriting the new generic types with those from `$this`,
and the type builder was discarding generic parameters for `static` and `self` types. Both have been corrected.

closes #266

Signed-off-by: azjezz <azjezz@protonmail.com>