feat(analyzer): infer template types from default parameter values
This commit adds support for inferring generic template types from the default values of optional parameters in function and method calls.

Previously, if an argument for a generic parameter was omitted, its corresponding template would remain unresolved.

The analyzer now inspects the default value's type to correctly infer the template for the specific invocation.

This also includes the necessary improvements to infer types for closures and arrow functions when they are used as default values.

related to #266

Signed-off-by: azjezz <azjezz@protonmail.com>