fix(analyzer): correctly infer type of `::class` on generic object
Previously, the type of `$variable::class` was inferred as `mixed` if `$variable` was of the generic object type.

This incorrectly treated the magic class constant like a regular class constant that would be unknown on a generic object.

The analyzer now correctly recognizes that `::class` always exists on any object and resolves its type to `class-string`.

closes #271

Signed-off-by: azjezz <azjezz@protonmail.com>