refactor(syntax, span): remove `#[repr(C)]` from AST nodes
The `#[repr(C)]` attribute was originally added to AST and span structs for potential C foreign function interface
(FFI) compatibility.

As there are no current or immediate plans for FFI with C, this attribute is unnecessary. Removing it simplifies the
code and gives the Rust compiler more freedom to optimize the memory layout of these structs for better performance.

Signed-off-by: azjezz <azjezz@protonmail.com>