fix(analyzer): correctly handle `readonly` property writes
This commit aligns the analyzer's behavior with the PHP engine's rules for readonly properties.

A `public readonly` property is now correctly treated as having `protected(set)` visibility.

This means the analyzer will now correctly flag any attempt to write to such a property from outside the class hierarchy.

Additionally, a new check has been implemented to report an error when a `readonly` property is written to outside of the constructor of its declaring class or one of its child classes,
which is also a fatal error in PHP if the property has already been initialized.

Signed-off-by: azjezz <azjezz@protonmail.com>